moves more code around
